Extract items into their own classes
